backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
google-generativeai==0.8.3
httpx==0.27.0
pydantic==2.5.3
tavily-python==0.5.0
